refactor: move Pagination DTO to common module
